# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "DocServer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
...
        "/profiles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows authenticated users to search for other user profiles within the system.\n\nYou can filter the search using query parameters in the URL:\n*   `email`: Find profiles where the email address contains the provided text (case-insensitive). Example: `?email=test.com`\n*   `first_name`: Find profiles where the first name contains the provided text (case-insensitive). Example: `?first_name=jo`\n*   `last_name`: Find profiles where the last name contains the provided text (case-insensitive). Example: `?last_name=smi`\nYou can combine multiple filters. The search returns profiles that match *all* provided filters.\n\nResults are paginated to handle potentially large numbers of users:\n*   `page`: Specifies which page of results to retrieve (starts at 1). Default is 1. Example: `?page=2`\n*   `limit`: Specifies how many profiles to return per page. Default is 20, maximum is 100. Example: `?limit=50`\n\nExample combining filters and pagination: `/profiles?first_name=a\u0026page=1\u0026limit=10` (Find profiles with 'a' in the first name, show the first 10 results).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Search User Profiles",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user@example.com",
                        "description": "Filter profiles where email contains this text (case-insensitive).",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "John",
                        "description": "Filter profiles where first name contains this text (case-insensitive).",
                        "name": "first_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Doe",
                        "description": "Filter profiles where last name contains this text (case-insensitive).",
                        "name": "last_name",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number for results (starts at 1).",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "example": 20,
                        "description": "Number of profiles per page.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of profiles matching the search criteria, along with pagination details (total count, current page, limit).",
                        "schema": {
                            "$ref": "#/definitions/api.SearchProfilesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid query parameters. 'page' and 'limit' must be positive integers. 'limit' cannot exceed 100.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired. You need to be logged in to search profiles.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while searching for profiles.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/profiles/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the profile details (like first name, last name, email, creation date) for the user who is currently logged in.\n\nThink of this as your \"My Account\" page data. To use this endpoint, you must first authenticate (log in) to get an access token.\nThe server uses the access token you provide in the request header to figure out who you are and fetch your specific profile information from the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Get Your Own Profile",
                "responses": {
                    "200": {
                        "description": "Your profile details were successfully retrieved. The response body contains your profile information (excluding sensitive data like the password hash).",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired. You might need to log in again.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found: The server couldn't find a profile associated with your access token. This is unusual if your token is valid.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server side (e.g., a database connection issue or a problem reading your user ID from the token context).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows the currently logged-in user to update their own profile information.\n\nYou can change your `first_name`, `last_name`, and any custom `extra` data associated with your profile.\n**Important:** You *cannot* change your email address or password using this endpoint. Password changes typically have a separate, more secure process (like a password reset flow).\nYou need to provide your current access token for authentication. The request body should contain the fields you want to update in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Update Your Own Profile",
                "parameters": [
                    {
                        "description": "The profile fields you want to update. 'first_name' and 'last_name' are required.",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Your profile was successfully updated. The response body contains the complete, updated profile.",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request: The data you sent in the request body is invalid. This could be due to missing required fields ('first_name', 'last_name') or incorrect JSON formatting.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired. You need to be logged in to update your profile.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found: The server couldn't find your profile based on your access token.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while trying to update your profile (e.g., a database error).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently deletes the account and profile data for the currently logged-in user.\n\n**WARNING: This action is irreversible!** Once you delete your account, all your data associated with it (profile, documents you own, etc.) will be removed.\n*(Developer Note: Full cascading delete logic, like removing shared document access, might still be under development. Currently, it primarily removes the main profile record.)*\nYou must provide your valid access token to authorize this action.",
                "tags": [
                    "Profiles"
                ],
                "summary": "Delete Your Own Profile",
                "responses": {
                    "204": {
                        "description": "Account Successfully Deleted. No content is returned in the response body because the resource (your profile) no longer exists."
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired. You need to be logged in to delete your account.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found: The server couldn't find your profile based on your access token.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while trying to delete your account (e.g., a database error).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.SearchProfilesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Profile"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.UpdateProfileRequest": {
            "type": "object",
            "required": [
                "first_name",
                "last_name"
            ],
            "properties": {
                "extra": {},
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Profile": {
            "type": "object",
            "properties": {
                "creation_date": {
                    "description": "UTC",
                    "type": "string"
                },
                "email": {
                    "description": "Unique, used for login",
                    "type": "string"
                },
                "extra": {
                    "description": "User-defined data"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID (UUID, dashless)",
                    "type": "string"
                },
                "last_modified_date": {
                    "description": "UTC",
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "utils.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}